requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
langchain>=0.0.300
python-multipart>=0.0.6
watchdog>=3.0.0
